pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-taggit>=6.1.0",
    "djangorestframework>=3.16.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
]
